Detect when a submodule repo can't be fetched from.
In my case this was caused by a refs/remote/origin/HEAD ref referring
to origin/master, after origin/master was renamed to origin/main.

More generally, this will print a louder and more explicit error if
go-git is refusing to fetch for some reason other than the ref not
existing.

https://github.com/go-git/go-git/pull/165